// src/enums/operationsTypes/operationType.enum.ts
export enum OPERATION_TYPE {
    CALCULATION = 'calculation',
    CLASSIFICATION = 'classification',
    COMPARISON = 'comparison',
    EXPORT = 'export',
    EXTRACTION = 'extraction',
    GENERATION = 'generation',
    IMPORT = 'import',
    SENTIMENT = 'sentiment',
    SUMMARIZATION = 'summarization',
    TRANSFORMATION = 'transformation',
    TRANSLATION = 'translation',
}